Deepend release 3.8
- For the Ada95, Ada2005, and Ada 2012 versions, the generic
Subpool_Allocators package adds the Default_Value as a generic
formal object, rather than as a function of the package. In addition,
the allocation_type can now be an unconstrained type, so it is
possible to allocate objects such as variable length Strings or
objects with immutable discriminants to a specific subpool.
It is still not possible to allocate objects of a limited type using
this approach, but for that, one can always use the Ada "new" keyword
syntax to allocate objects of limited types to the pool, but these
allocations will always go to the default subpool, unless the
the Ada 2012 subpool allocator syntax is used, in which case the
allocation goes to the specified subpool.